Split out the managers into a separate file.
* And the same with their tests.
* I also got mad docstrings, yo.
